requirements.txt
django==1.4
lxml
psycopg2>=2.3
supervisor==3.0a12
xlrd
# Unsure which version satisfied minimum reqs
PIL

django-haystack==1.2.7
django-piston==0.2.2
django-tinymce
python-memcached

# The way we use pysolr introduces an extra dependency:
pysolr==2.0.15
BeautifulSoup==3.2.1

# For memcached on Heroku we need a caching back-end that does authentication:
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4

# -e src/facebook-python-sdk-322930c
# -e .
# --extra-index-url=http://dist.pinaxproject.com/dev/ -e ../myplants

# For storing images on S3.
django-storages==1.1.4
boto==2.2.2

# For deployment to Heroku.
gunicorn
newrelic

# For our Heroku Scheduler tasks that scan S3
s3cmd

# For testing.
coverage
unittest2
selenium
